graduates changed to users and everything ammended accordingly
