+++ server/routes/users.js
@@ -1,9 +1,10 @@
 const express = require('express')
 const router = express.Router()
-const db = require('../db/graduates')
+const db = require('../db/users')
 
 router.get('/', (req, res) => {
-  db.getGraduates()
+  const status = 'true'
+  db.getGraduates(status)
     .then(graduates => {
       res.json(graduates)
     })
